refactor: extractor turned into OOP
