--- src/tab/scraper/extractor/ChatbotExtractor.js
@@ -1,11 +1,0 @@
-import {Extractor} from "./Extractor";
-
-export class ChatbotExtractor extends Extractor {
-  constructor(domain) {
-    super(domain);
-  }
-
-  extractSources() {
-    throw new Error("Not implemented");
-  }
-}
