+++ src/tab/scraper/extractor/extract.js
@@ -1,5 +1,3 @@
-import extractPages from "./extractPages";
-import extractMetadata from "./extractMetadata";
 import converter from "../../formatter/formatMarkdown";
 import {formatFilename} from "../../formatter/formatText";
 import {dynamicCall} from "../../../common/utils";
@@ -7,9 +5,37 @@ import appInfos from "../../../infos.json";
 
 export async function extract(domain) {
   try {
-    const markdownContent = await dynamicCall(extractPages, `extract${domain.name}Page`, converter[`formatMarkdown`]);
-    const metadata = dynamicCall(extractMetadata, `extract${domain.name}Metadata`);
-    const fileName = formatFilename(metadata.title, metadata.source); //???
+    console.log(domain)
+    let module;
+    switch (domain.name) {
+      case "PhindSearch":
+        module = await import(`./ExtractorPhindSearch`);
+        break;
+      case "PhindAgent":
+        module = await import(`./ExtractorPhindAgent`);
+        break;
+      case "Perplexity":
+        module = await import(`./ExtractorPerplexity`);
+        break;
+      case "MaxAIGoogle":
+        module = await import(`./ExtractorMaxAIGoogle`);
+        break;
+      default:
+        module = await import(`./ExtractorArbitraryPage`);
+    }
+    // let extractor = await import(`./Extractor${domain.name}`);
+    console.log(module);
+
+    let extractor = new module.default(domain);
+    console.log(extractor);
+
+
+    const markdownContent = await extractor.extractPage(converter[`formatMarkdown`]) //await dynamicCall(extractPages, `extract${domain.name}Page`, converter[`formatMarkdown`]);
+    console.log(markdownContent)
+    const metadata = await extractor.extractMetadata() //await dynamicCall(extractMetadata, `extract${domain.name}Metadata`);
+    console.log(metadata)
+    const fileName = formatFilename(metadata.title, metadata.source);
+    console.log(fileName)
     return {markdownContent, title: metadata.title, fileName};
   }
   catch (e) {
